LayoutTests/networkinformation/add-listener-from-callback-expected.txt

Tests that adding a new event listener from a callback works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof connection.bandwidth is "number"
PASS typeof connection.metered is "boolean"
PASS typeof connection.bandwidth is "number"
PASS typeof connection.metered is "boolean"
PASS typeof connection.bandwidth is "number"
PASS typeof connection.metered is "boolean"
PASS successfullyParsed is true

TEST COMPLETE
